Created an src folder as well.
